added ijsonLoad() for faster loading; fixed bug fixedInterval int

+++ netpyne/cell.py
@@ -1349,7 +1349,7 @@ class PointCell (Cell):
                         end = pulse['end']
 
                         # fixed interval of duration (1 - noise)*interval 
-                        fixedInterval = np.full(((1+1.5*noise)*(end-start)/interval), [(1.0-noise)*interval])  # generate 1+0.5*noise spikes to account for noise
+                        fixedInterval = np.full(int(((1+1.5*noise)*(end-start)/interval)), [(1.0-noise)*interval])  # generate 1+0.5*noise spikes to account for noise
                         numSpks = len(fixedInterval)
 
                         # randomize the first spike so on average it occurs at start + noise*interval
